backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
spotipy==2.23.0
youtube-search-python==1.6.6
SpeechRecognition==3.10.0
pydub==0.25.1
httpx==0.25.0
redis==5.0.1